+++ src/pipx/main.py
@@ -238,7 +238,9 @@ def run_pipx_command(args: argparse.Namespace) -> ExitCode:  # noqa: C901
             force=args.force,
         )
     elif args.command == "list":
-        return commands.list_packages(venv_container, args.include_injected)
+        return commands.list_packages(
+            venv_container, args.include_injected, args.format
+        )
     elif args.command == "uninstall":
         return commands.uninstall(venv_dir, constants.LOCAL_BIN_DIR, verbose)
     elif args.command == "uninstall-all":
@@ -494,6 +496,12 @@ def _add_list(subparsers) -> None:
         "--include-injected",
         action="store_true",
         help="Show packages injected into the main app's environment",
+    )
+    p.add_argument(
+        "--format",
+        choices=["human", "freeze", "json"],
+        default="human",
+        help="Control the format output in a machine readable manner",
     )
     p.add_argument("--verbose", action="store_true")